feat(settings): introduce new settings primitives and refactor imports
